feat(add-ask-cmd): add first interation of asking command
