+++ cargo-projects/vivs/src/commands.rs
@@ -7,7 +7,7 @@ use crate::data_chunk::DataChunkError;
 use crate::parser::Parser;
 use crate::utils::{unknown_cmd_err, NO_CMD_ERR};
 use crate::{ClusterInstanceConfig, Connection, DataStore, GenericResult};
-use asking::Ask;
+// use asking::Ask;
 use core::str;
 use delete::Delete;
 use get::Get;
@@ -31,7 +31,7 @@ pub enum Command {
     Set(Set),
     Delete(Delete),
     Ttl(Ttl),
-    Ask(Ask),
+    // Ask(Ask),
     Unknown(String),
     None,
 }
@@ -67,47 +67,6 @@ impl From<&str> for ParseCommandErr {
     }
 }
 
-pub trait AskCommand {
-    async fn check_ask(&self, key: &str, conn: &mut Connection) -> Option<(u16, String)> {
-        // check if config exists, we'll most likely need to store it in memory to avoid constant IO (?)
-        let own_addr = conn.own_addr().unwrap().to_string();
-
-        let port = own_addr.split(":").nth(1).unwrap_or_default();
-        let current_dir = current_dir().unwrap();
-        let node_config = format!("{}/{}.toml", current_dir.display(), port);
-
-        // When no cluster (<port>.conf) file is found,
-        // We can assume that Vivs is not running in the cluster mode.
-        // Therefore normal processing of incoming command should take place.
-        let Ok(file_contents) = fs::read(node_config).await else {
-            return None;
-        };
-
-        let file_contents = str::from_utf8(&file_contents[0..]).unwrap();
-        let nodes = toml::from_str::<ClusterInstanceConfig>(file_contents).unwrap();
-
-        // Work out a cell / hash slot
-        const X25: crc::Crc<u16> = crc::Crc::<u16>::new(&crc::CRC_16_IBM_SDLC);
-        let key_hash = X25.checksum(key.as_bytes()) % 16384;
-
-        // Iterate over all current nodes in the cluster
-        for (ip, config) in nodes {
-            let cell_range = config.position.0..config.position.1;
-
-            let is_in_range = cell_range.contains(&key_hash.into());
-            if own_addr == ip && is_in_range {
-                return None;
-            }
-
-            if is_in_range {
-                return Some((key_hash, ip));
-            }
-        }
-
-        None
-    }
-}
-
 pub trait CommonCommand {
     fn parse(data: Parser) -> Self;
     fn respond(
@@ -118,11 +77,11 @@ pub trait CommonCommand {
 }
 
 impl Command {
-    fn process_location(a: Command) -> Command {
-        Command::Ask(Ask {
-            command: Box::new(a),
-        })
-    }
+    // fn process_location(a: Command) -> Command {
+    //     Command::Ask(Ask {
+    //         command: Box::new(a),
+    //     })
+    // }
 
     pub fn parse_cmd(mut data_chunk: Parser) -> Result<Command, ParseCommandErr> {
         // The iterator should contain all the necessary commands and values e.g. [SET, key, value]
@@ -136,7 +95,7 @@ impl Command {
         let command = match &command[..] {
             PING_CMD => Command::Ping(Ping::parse(data_chunk)),
             GET_CMD => Command::Get(Get::parse(data_chunk)),
-            SET_CMD => Self::process_location(Command::Set(Set::parse(data_chunk))),
+            SET_CMD => Command::Set(Set::parse(data_chunk)),
             DELETE_CMD => Command::Delete(Delete::parse(data_chunk)),
             TTL_CMD => Command::Ttl(Ttl::parse(data_chunk)),
             "" => Command::None,
@@ -153,7 +112,7 @@ impl Command {
             Command::Set(command) => command.respond(conn, db).await,
             Command::Delete(command) => command.respond(conn, db).await,
             Command::Ttl(command) => command.respond(conn, db).await,
-            Command::Ask(command) => command.respond(conn).await,
+            // Command::Ask(command) => command.respond(conn).await,
             Command::None => {
                 conn.write_error(NO_CMD_ERR.as_bytes()).await?;
                 Ok(())
